Setup: db connection, schema, admin user (#62)
* setup cmd  begin

* reading with readfile

* split setup function

* setup schema

* improve error message

* reconnect

* setup users

* rm user to console

+++ packages/saltcorn-data/db/reset_schema.js
@@ -1,14 +1,16 @@
 const db = require(".");
 
 //https://stackoverflow.com/a/21247009
-const reset = async () => {
-  await db.query(`
+const reset = async dontDrop => {
+  if (!dontDrop) {
+    await db.query(`
     DROP SCHEMA public CASCADE;
     CREATE SCHEMA public;
     GRANT ALL ON SCHEMA public TO postgres;
     GRANT ALL ON SCHEMA public TO public;
     COMMENT ON SCHEMA public IS 'standard public schema';
   `);
+  }
 
   await db.query(`
     CREATE TABLE tables
